requirements.txt
# -----------WHAM---------
# chumpy @ git+https://github.com/mattloper/chumpy
# numpy==1.22.3
# yacs
# joblib
# scikit-image
# opencv-python
# imageio[ffmpeg]
# matplotlib
tensorboard
# smplx
progress
einops
mmcv==1.3.9
# timm==0.4.9
munkres
xtcocotools>=1.8
# loguru
# setuptools==59.5.0
# tqdm
# ultralytics
# gdown==4.6.0

#-------TokenHMR---------
# numpy==1.23.1
# pytorch-lightning==1.8.1
# smplx==0.1.28
# pyrender==0.1.45
# opencv-python
# yacs
# scikit-image
# einops
# timm
# dill
# chumpy
# pandas
# rich
# hydra-core
# hydra-submitit-launcher
# hydra-colorlog
# pyrootutils
# webdataset
transformers
# flatten_dict
git+https://github.com/nghorbani/body_visualizer
# setuptools==65.0.0

#-------CameraHMR---------
# pytorch-lightning==2.0.2
# smplx==0.1.28
# pyrender==0.1.45
# opencv-python==4.10.0.84
# yacs==0.1.8
# scikit-image==0.21.0
# einops==0.8.0
# timm==1.0.12
# pandas==2.1.1
# loguru==0.7.3
# numpy==1.23.5
# pyrootutils==1.0.4
# rich==14.0.0
# hydra-colorlog==1.2.0
albumentations==1.3.1
# chumpy
# flatten_dict
jpeg4py
# joblib

#-------hmr2.0---------
# pytorch-lightning
# smplx==0.1.28
# pyrender
# opencv-python
# yacs
# scikit-image
# einops
# timm
# dill
# pandas
# git+https://github.com/facebookresearch/detectron2
# git+https://github.com/mattloper/chumpy
# rich
# hydra-core
# hydra-submitit-launcher
# hydra-colorlog
# pyrootutils
# webdataset


# # -------PromptHMR---------
# evo==1.29.0
# # yacs==0.1.8
# # gdown==5.2.0
# # einops==0.8.0
# open3d==0.19.0
# # loguru==0.7.2
# plyfile==1.0.3
# # pyrender==0.1.45
# omegaconf==2.3.0
# keyboard==0.13.5

# pycocotools==2.0.8
# # timm==0.9.0
# # ultralytics==8.3.80
# supervision==0.25.1
# # scikit-image==0.24.0
# segment-anything==1.0
# pytorch-minimize==0.0.2
# # pytorch-lightning==2.4.0
# open_clip_torch
# viser==0.2.23
# pyequilib==0.3.0
# mmcv-lite==2.2.0
# mmengine==0.10.4

# # opencv-python-headless==4.10.0.84
# # imageio[ffmpeg]
# ffmpeg==1.4
# openpyxl==3.1.5
# filterpy==1.4.5
# # smplx==0.1.28
# # fvcore
# pyliblzfse
# smplcodec
# smpl-rs==0.6.1
# gloss-rs==0.6.0
# # git+https://github.com/mattloper/chumpy@9b045ff5d6588a24a0bab52c83f032e2ba433e17
# # git+https://github.com/facebookresearch/pytorch3d.git@stable

#-------Conflict---------
git+https://github.com/facebookresearch/detectron2.git
ultralytics
gdown==5.2.0
imageio[ffmpeg]
matplotlib
numpy==1.23.5 
tqdm
smplx==0.1.28
timm==1.0.12
setuptools==65.0.0  
chumpy
pytorch-lightning==2.4.0
pyrender==0.1.45
opencv-python-headless==4.10.0.84
yacs==0.1.8
scikit-image==0.24.0
einops==0.8.0
pandas
loguru==0.7.3
pyrootutils==1.0.4
rich==14.0.0
hydra-core
hydra-submitit-launcher
hydra-colorlog
flatten_dict
joblib
dill
webdataset

